Added SignUp mockup
