+++ .eslintrc.json
@@ -14,12 +14,12 @@
       "aspects": [ "noHref", "invalidHref", "preferButton" ]
     }],
     "jsx-a11y/label-has-for": [ 2, {
-      "components": [ "Label" ],
-      "required": {
-          "some": [ "nesting", "id" ]
-      },
-      "allowChildren": false
-  }]
+        "components": [ "Label" ],
+        "required": {
+            "some": [ "nesting", "id" ]
+        },
+        "allowChildren": false
+    }]
   }
 }
