+++ app/auth/components/AuthForm/index.js
@@ -1,0 +1,3 @@
+import AuthForm from './AuthForm';
+
+export default AuthForm;
