+++ app/auth/components/AuthForm/styles.scss
@@ -1,4 +1,4 @@
-.login-container {
+.auth-form-container {
   margin-top: 7em;
   .form-labels {
     text-align: left;
