+++ app/auth/components/Main.js
@@ -1,14 +1,13 @@
 import React from 'react';
 import { Switch, Route, Redirect } from 'react-router-dom';
 
-import SignUp from './SignUp';
-import Login from './Login';
+import AuthForm from './AuthForm';
 
 const Main = () => (
   <Switch>
     <Route exact path="/" render={() => (<Redirect to="/login" />)} />
-    <Route exact path="/signup" component={SignUp} />
-    <Route exact path="/login" component={Login} />
+    <Route exact path="/signup" component={() => <AuthForm isSignUp />} />
+    <Route exact path="/login" component={AuthForm} />
   </Switch>
 );
