+++ app/auth/components/Navbar.scss
@@ -1,0 +1,3 @@
+.navbar {
+  border-top: 0;
+}
