--- app/auth/components/SignUp.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import { Form, Grid } from 'semantic-ui-react';
-
-
-const SignUp = () => (
-  <Grid>
-    <Grid.Row>
-      <Form>
-      Sign Up
-      </Form>
-    </Grid.Row>
-  </Grid>
-);
-
-export default SignUp;
